requirements.txt
streamlit>=1.44.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.20.0
yfinance>=0.2.54
scikit-learn>=1.4.0
vaderSentiment>=3.3.2
feedparser>=6.0.11
beautifulsoup4>=4.12.3
openai>=1.72.0
Pillow>=10.4.0
